Feat/support omv ide

See merge request ae_group/openmv!3

+++ lib/imlib/imlib.h
@@ -206,7 +206,7 @@ typedef struct color_thresholds_list_lnk_data {
         __typeof__ (pixel0) _pixel0 = (pixel0);          \
         __typeof__ (pixel1) _pixel1 = (pixel1);          \
         __typeof__ (threshold) _threshold = (threshold); \
-        (abs(_pixel0 - _pixel1) <= _threshold);          \
+        (abs((int32_t) _pixel0 - (int32_t) _pixel1) <= _threshold); \
     })
 
 #define COLOR_BOUND_GRAYSCALE(pixel0, pixel1, threshold) \
@@ -351,7 +351,7 @@ extern const int8_t lab_table[196608 / 2];
 #define COLOR_BINARY_TO_RGB565(pixel)           COLOR_YUV_TO_RGB565(((pixel) ? 127 : -128), 0, 0)
 #define COLOR_RGB565_TO_BINARY(pixel)           (COLOR_RGB565_TO_Y(pixel) > (((COLOR_Y_MAX - COLOR_Y_MIN) / 2) + COLOR_Y_MIN))
 #define COLOR_RGB565_TO_GRAYSCALE(pixel)        COLOR_RGB565_TO_Y(pixel)
-#define COLOR_GRAYSCALE_TO_BINARY(pixel)        ((pixel) > \
+#define COLOR_GRAYSCALE_TO_BINARY(pixel)        (((int) (pixel)) > \
                                                  (((COLOR_GRAYSCALE_MAX - COLOR_GRAYSCALE_MIN) / 2) + COLOR_GRAYSCALE_MIN))
 #define COLOR_GRAYSCALE_TO_RGB565(pixel)        COLOR_YUV_TO_RGB565(((pixel) - 128), 0, 0)
